Configuración de la página de gestión de alumnos, tango GridView como buscador Implementadas las funciones de sincronización de los alumnos Implementadas vistas de actualización y creación de alumnos Implementación de reglas y funciones para corregir el DNI/NIE en caso de que sea incorrecto Implementado la búsqueda por el nombre completo de un alumno Corrección de la creación de las siglas de los cursos

+++ controllers/AlumnosController.php
@@ -6,7 +6,7 @@ use Yii;
 use app\models\Alumnos;
 use app\models\AlumnosSearch;
 use app\models\Cursan;
-use yii\data\ActiveDataProvider;
+use app\models\Portatiles;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
@@ -44,6 +44,7 @@ class AlumnosController extends Controller {
             return $this->goHome();
         }
 
+        Portatiles::sincronizarPortatiles();
         Alumnos::sincronizarAlumnos();
         Cursan::sincronizarCursan();
 
@@ -51,54 +52,48 @@ class AlumnosController extends Controller {
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
         $model = new Alumnos();
 
-        return $this->render('index', [
-            'dataProvider' => $dataProvider,
-            'searchModel' => $searchModel,
-            'model' => $model
-        ]);
-
-    }
-
-    /**
-     * Displays a single Alumnos model.
-     * @param int $id_alumno Id Alumno
-     * @return string
-     * @throws NotFoundHttpException if the model cannot be found
-     */
-    public function actionView($id_alumno) {
-
-        if(Yii::$app->user->isGuest) {
-            return $this->goHome();
-        }
-
-        return $this->render('view', [
-            'model' => $this->findModel($id_alumno),
-        ]);
-
-    }
-
-    /**
-     * Creates a new Alumnos model.
-     * If creation is successful, the browser will be redirected to the 'view' page.
-     * @return string|\yii\web\Response
-     */
-    public function actionCreate() {
-
-        if(Yii::$app->user->isGuest) {
-            return $this->goHome();
-        }
-
-        $model = new Alumnos();
-
         if ($this->request->isPost) {
-            if ($model->load($this->request->post()) && $model->save()) {
-                return $this->redirect(['view', 'id_alumno' => $model->id_alumno]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+
+                if ($model->estado_matricula == 'Matriculado') {
+
+                    $cursoManana = Yii::$app->request->post('cursoManana');
+                    $cursoTarde = Yii::$app->request->post('cursoTarde');
+
+                    if (!empty($cursoManana) || !empty($cursoTarde)) {
+
+                        $cursos = [$cursoManana, $cursoTarde];
+
+                        foreach ($cursos as $curso) {
+                            if (!empty($curso)) {
+                                $modelCursa = new Cursan();
+                                $modelCursa->id_curso = $curso;
+                                $modelCursa->id_alumno = $model->id_alumno;
+                                $modelCursa->curso_academico = Cursan::getCursoActual();
+                                $modelCursa->save();
+                            }
+                        }
+
+                    } else {
+                        $model->id_portatil = null;
+                        $model->save();    
+                    }
+
+                } else {
+                    $model->id_portatil = null;
+                    $model->save();
+                }
+
+                return $this->redirect(['index']);
+
             }
         } else {
             $model->loadDefaultValues();
         }
 
-        return $this->render('create', [
+        return $this->render('index', [
+            'dataProvider' => $dataProvider,
+            'searchModel' => $searchModel,
             'model' => $model,
         ]);
 
@@ -118,13 +113,73 @@ class AlumnosController extends Controller {
         }
 
         $model = $this->findModel($id_alumno);
+        $cursoActualManana = $model->cursoManana;
+        $cursoActualTarde = $model->cursoTarde;
 
-        if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id_alumno' => $model->id_alumno]);
+        if ($this->request->isPost) {
+            
+            if ($model->load($this->request->post()) && $model->save()) {
+
+                if ($model->estado_matricula == 'Matriculado') {
+
+                    $cursoManana = Yii::$app->request->post('cursoManana');
+                    $cursoTarde = Yii::$app->request->post('cursoTarde');
+
+                    if (!empty($cursoManana) || !empty($cursoTarde)) {
+
+                        if (!empty($cursoManana)) {
+                            if (($cursoActualManana) ? $cursoManana != $cursoActualManana->id_curso : true) {
+                                $modelCursa = new Cursan();
+                                $modelCursa->id_curso = $cursoManana;
+                                $modelCursa->id_alumno = $model->id_alumno;
+                                $modelCursa->curso_academico = Cursan::getCursoActual();
+                                $modelCursa->save();
+                                if ($cursoActualManana) {
+                                    Cursan::find()->where(['id_curso' => $cursoActualManana->id_curso])->andWhere(['id_alumno' => $model->id_alumno])->one()->delete();
+                                }
+                            }
+                        } else if ($cursoActualManana) {
+                            Cursan::find()->where(['id_curso' => $cursoActualManana->id_curso])->andWhere(['id_alumno' => $model->id_alumno])->one()->delete();
+                        }
+
+                        if (!empty($cursoTarde)) {
+                            if (($cursoActualTarde) ? $cursoTarde != $cursoActualTarde->id_curso : true) {
+                                $modelCursa = new Cursan();
+                                $modelCursa->id_curso = $cursoTarde;
+                                $modelCursa->id_alumno = $model->id_alumno;
+                                $modelCursa->curso_academico = Cursan::getCursoActual();
+                                $modelCursa->save();
+                                if ($cursoActualTarde) {
+                                    Cursan::find()->where(['id_curso' => $cursoActualTarde->id_curso])->andWhere(['id_alumno' => $model->id_alumno])->one()->delete();
+                                }
+                            }
+                        } else if ($cursoActualTarde) {
+                            Cursan::find()->where(['id_curso' => $cursoActualTarde->id_curso])->andWhere(['id_alumno' => $model->id_alumno])->one()->delete();
+                        }
+
+                    } else {
+                        $model->id_portatil = null;
+                        $model->save();    
+                    }
+
+                } else {
+                    $model->id_portatil = null;
+                    $model->save();
+                }
+
+
+                return $this->redirect(['index']);
+
+            }
+
+        } else {
+            $model->loadDefaultValues();
         }
 
         return $this->render('update', [
             'model' => $model,
+            'cursoActualManana' => $cursoActualManana,
+            'cursoActualTarde' => $cursoActualTarde,
         ]);
 
     }
@@ -147,6 +202,16 @@ class AlumnosController extends Controller {
      * @throws NotFoundHttpException if the model cannot be found
      */
     public function actionDelete($id_alumno) {
+
+        if(Yii::$app->user->isGuest) {
+            return $this->goHome();
+        }
+
+        $aplicaciones = Cursan::find()->where(['id_alumno' => $id_alumno])->all();
+
+        foreach ($aplicaciones as $aplicacion) {
+            $aplicacion->delete();
+        }
 
         $this->findModel($id_alumno)->delete();
 
@@ -171,7 +236,7 @@ class AlumnosController extends Controller {
             return $model;
         }
 
-        throw new NotFoundHttpException('The requested page does not exist.');
+        throw new NotFoundHttpException('❌ El alumno/a no existe');
 
     }
